% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lvd.imports.R
\name{build.lvd}
\alias{build.lvd}
\title{build.lvd}
\usage{
build.lvd(load, displ, time = NULL)
}
\arguments{
\item{load}{A vector of numeric values specifying the load.}

\item{displ}{A vector of numeric values specifying the displacement.}

\item{time}{An optional vector of numeric values specifying the time.}
}
\value{
An object of class lvd (load versus displacement)
}
\description{
Builds an object of class .lvd from specified vectors
}
\examples{

}
