% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lvd.plots.R
\name{lvd.plot}
\alias{lvd.plot}
\title{lvd.plot}
\usage{
lvd.plot(x, auc = TRUE)
}
\arguments{
\item{x}{An object of class 'lvd'}

\item{auc}{A logical indicating whether the area under curve should be filled}
}
\value{
A ggplot2 object
}
\description{
Plot load vs. displacement curve
}
\examples{

}
