% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lvd.imports.R
\docType{class}
\name{lvd}
\alias{lvd}
\title{The title for my S4 class}
\description{
Some details about this class and my plans for it in the body.
}
\section{Slots}{

\describe{
\item{\code{load}}{A vector of numeric values specifying the load.}

\item{\code{displacement}}{A vector of numeric values specifying the displacement.}

\item{\code{time}}{An optional vector of numeric values specifying the time.}
}}

