% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lvd.conversions.R
\name{lvd.to.strain}
\alias{lvd.to.strain}
\title{lvd.to.str}
\usage{
lvd.to.strain(x, lng, unit = "mm")
}
\arguments{
\item{x}{An object of class 'lvd', or a numeric vector of displacement values}

\item{lng}{A numeric indicating the initial specimen length}

\item{unit}{The unit in which the measure was taken. Default is mm.}
}
\value{
A numeric vector for strain
}
\description{
Convert load data from an lvd object into strain
}
\examples{

}
