% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lvd.conversions.R
\name{lvd.to.stress}
\alias{lvd.to.stress}
\title{lvd.to.stress}
\usage{
lvd.to.stress(x, area, unit = "mm2")
}
\arguments{
\item{x}{A vector of load data or an object of class 'lvd'}

\item{area}{A numeric indicating the initial specimen cross-section area}

\item{unit}{The unit in which the measure was taken. Default is square mm ("mm2").}
}
\value{
A numeric vector for stress
}
\description{
Convert load data from an lvd object into stress
}
\examples{

}
