% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lvd.imports.R
\name{read.lvm}
\alias{read.lvm}
\title{read.lvm}
\usage{
read.lvm(file, abs = TRUE)
}
\arguments{
\item{file}{An .lvm file}

\item{abs}{Logical indicating whether data should be read as exact (abs = FALSE) or absolute values (abs = TRUE).}
}
\value{
An object of class lvd (load versus displacement)
}
\description{
Loads a .lvm file (from Lucas' mechanical tester FLS2) into memory
}
\examples{

}
