% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/description.stiffr.R
\docType{package}
\name{stiffr}
\alias{stiffr}
\title{stiffr: An R package for converting load vs. displacement data into physical properties}
\description{
The R package 'stiffr' converts load vs. time or displacement data into physical properties (hardness, toughness, Young’s modulus, strength...)
It also allows to build load vs. displacement curves.
}
