% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lvd.conversions.R
\name{strength}
\alias{strength}
\title{strength}
\usage{
strength()
}
\arguments{
\item{x}{An object of class 'lvd'}
}
\value{
A numeric value for strength
}
\description{
Convert data from an lvd object into strength
}
\examples{

}
