% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lvd.conversions.R
\name{vickers.area}
\alias{vickers.area}
\title{vickers.area}
\usage{
vickers.area(d)
}
\arguments{
\item{d}{the average diagonal length of the indentation mark}
}
\value{
A numeric approcimation of the surface area of the indentation mark
}
\description{
Converts dimensions of indentation marks left by Vicker's test into surface area
}
\examples{

}
