% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lvd.conversions.R
\name{young.modulus}
\alias{young.modulus}
\title{young.modulus}
\usage{
young.modulus(
  x,
  lng,
  area,
  unit.area = "mm2",
  unit.load = "N",
  crop = FALSE,
  zero.out = FALSE
)
}
\arguments{
\item{x}{An object of class 'lvd'}

\item{lng}{A numeric indicating the initial specimen length}

\item{area}{A numeric indicating the initial specimen cross-section area}

\item{unit.area}{The unit in which the cross area was taken. Default is square mm ("mm2")}

\item{unit.load}{The unit in which the load was measured. Default is N}

\item{crop}{Logical indicating whether the lvd.crop function should be launched before the computation starts}

\item{zero.out}{Logical indicating whether the stress-strain curve should be "zero-outed" before the conversion}
}
\value{
A numeric value for Young's elasticity modulus
}
\description{
Compute Young's modulus from an lvd object
}
\examples{

}
